requirements.txt
Flask 
numpy 
pillow
